immunochain-core/requirements.txt
aniso8601==6.0.0
bcrypt==3.1.7
cassandra-driver==3.17.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorlog==4.0.2
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT==0.3.2
Flask-JWT-Extended==3.20.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
passlib==1.7.1
pika==1.0.0
protobuf==3.7.1
psycopg2==2.8.2
pyasn1==0.4.5
pycparser==2.19
PyJWT==1.7.1
pymongo==3.7.2
pypng==0.0.20
PyQRCode==1.2.1
PySocks==1.7.1
pytz==2018.9
PyYAML==5.1
pyzmq==18.0.1
requests==2.21.0
rsa==4.0
sawtooth-sdk==1.1.4.post1
sawtooth-signing==1.1.4.post1
secp256k1==0.13.2
shortuuid==0.5.0
six==1.12.0
SQLAlchemy==1.3.5
toml==0.10.0
twilio==6.31.0
urllib3==1.24.1
Werkzeug==0.15.0
python-dateutil==2.8.1
